requirements.txt
scipy==1.4.1
torch==1.10.0
torchvision==0.11.1
tqdm
pandas
matplotlib
easydict==1.9
Pillow==7.1.2